@@ -32,7 +32,7 @@
         "test:integ:cov": "nyc --report-dir ./coverage/integration yarn run test:integ"
     },
     "dependencies": {
-        "ini": "^3.0.1",
+        "ini": "^4.0.0",
         "inversify": "^6.0.1",
         "node-fetch-commonjs": "^3.2.4",
         "reflect-metadata": "^0.1.13"
@@ -42,20 +42,20 @@
         "@sinonjs/fake-timers": "^10.0.2",
         "@types/ini": "^1.3.31",
         "@types/mocha": "^10.0.1",
-        "@types/node": "^18.13.0",
+        "@types/node": "^18.15.11",
         "@types/sinonjs__fake-timers": "^8.1.2",
         "@types/tmp": "^0.2.3",
-        "@types/uuid": "^9.0.0",
-        "eslint": "^8.34.0",
+        "@types/uuid": "^9.0.1",
+        "eslint": "^8.38.0",
         "mocha": "^10.2.0",
         "nyc": "^15.1.0",
-        "prettier": "^2.8.4",
+        "prettier": "^2.8.7",
         "tmp-promise": "^3.0.3",
         "ts-loader": "^9.4.2",
         "ts-mocha": "^10.0.0",
         "ts-mockito": "^2.6.1",
         "ts-node": "^10.9.1",
-        "typescript": "^4.9.5",
+        "typescript": "^5.0.4",
         "uuid": "^9.0.0"
     },
     "nyc": {
