@@ -632,55 +665,58 @@
         "test": "yarn run test:lint && yarn run test:unit",
         "clean": "rm -rf node_modules out .vscode-test",
         "generate-notice": "../../scripts/generate_notice.sh",
-        "ts-node": "ts-node"
+        "generate-telemetry": "ts-node ./scripts/generateTelemetry.ts"
     },
     "dependencies": {
         "@databricks/databricks-sdk": "workspace:^",
         "@databricks/databricks-vscode-types": "workspace:^",
         "@vscode/debugadapter": "^1.59.0",
-        "@vscode/webview-ui-toolkit": "^1.2.1",
+        "@vscode/extension-telemetry": "^0.7.7",
+        "@vscode/webview-ui-toolkit": "^1.2.2",
         "ansi-to-html": "^0.7.2",
+        "bcryptjs": "^2.4.3",
         "triple-beam": "^1.3.0",
         "winston": "^3.8.2"
     },
     "devDependencies": {
         "@istanbuljs/nyc-config-typescript": "^1.0.2",
         "@sinonjs/fake-timers": "^10.0.2",
+        "@types/bcryptjs": "^2.4.2",
         "@types/fs-extra": "^11.0.1",
-        "@types/glob": "^8.0.1",
         "@types/mocha": "^10.0.1",
         "@types/mock-require": "^2.0.1",
-        "@types/node": "^18.13.0",
+        "@types/node": "^18.15.11",
         "@types/sinonjs__fake-timers": "^8.1.2",
         "@types/tmp": "^0.2.3",
         "@types/triple-beam": "^1.3.2",
         "@types/vscode": "^1.69.1",
-        "@types/yargs": "^17.0.22",
-        "@typescript-eslint/eslint-plugin": "^5.52.0",
-        "@typescript-eslint/parser": "^5.52.0",
-        "@vscode/test-electron": "^2.2.3",
-        "@wdio/cli": "^7.30.1",
-        "@wdio/local-runner": "^7.30.1",
-        "@wdio/mocha-framework": "^7.26.0",
-        "@wdio/spec-reporter": "^7.29.1",
-        "@wdio/types": "^7.26.0",
-        "esbuild": "^0.17.8",
-        "eslint": "^8.34.0",
-        "fs-extra": "^11.1.0",
-        "glob": "^8.1.0",
+        "@types/yargs": "^17.0.24",
+        "@typescript-eslint/eslint-plugin": "^5.58.0",
+        "@typescript-eslint/parser": "^5.58.0",
+        "@vscode/test-electron": "^2.3.0",
+        "@wdio/cli": "^8.8.1",
+        "@wdio/local-runner": "^8.8.1",
+        "@wdio/mocha-framework": "^8.8.0",
+        "@wdio/spec-reporter": "^8.8.0",
+        "@wdio/types": "^8.8.0",
+        "esbuild": "^0.17.16",
+        "eslint": "^8.38.0",
+        "fs-extra": "^11.1.1",
+        "glob": "^10.0.0",
         "mocha": "^10.2.0",
         "mock-require": "^3.0.3",
+        "node-fetch-commonjs": "^3.2.4",
         "nyc": "^15.1.0",
-        "prettier": "^2.8.4",
+        "prettier": "^2.8.7",
         "tmp-promise": "^3.0.3",
         "ts-mockito": "^2.6.1",
         "ts-node": "^10.9.1",
-        "typescript": "^4.9.5",
+        "typescript": "^5.0.4",
         "vsce": "^2.15.0",
-        "wdio-video-reporter": "^3.5.0",
-        "wdio-vscode-service": "^4.2.1",
+        "wdio-video-reporter": "^4.0.1",
+        "wdio-vscode-service": "patch:wdio-vscode-service@npm:5.0.0#../../.yarn/patches/wdio-vscode-service-npm-5.0.0-931a93aae2.patch",
         "winston": "^3.8.2",
-        "yargs": "^17.7.0"
+        "yargs": "^17.7.1"
     },
     "nyc": {
         "extends": "@istanbuljs/nyc-config-typescript",
