@@ -16,6 +16,7 @@ describe(__filename, () => {
 
     beforeEach(() => {
         connection = mock<ConnectionManager>();
+        when(connection.metadataServiceUrl).thenReturn("http://localhost:1234");
         cli = mock<CliWrapper>();
     });
 
@@ -64,6 +65,9 @@ describe(__filename, () => {
                     "profile"
                 )
             );
+            when(mockDbWorkspace.host).thenReturn(
+                Uri.parse("https://000000000000.00.azuredatabricks.net/")
+            );
 
             const mockSyncDestination = mock(SyncDestinationMapper);
             when(mockSyncDestination.localUri).thenReturn(
@@ -100,8 +104,10 @@ describe(__filename, () => {
                     BRICKS_ROOT: Uri.file("/path/to/local/workspace").fsPath,
                     BRICKS_UPSTREAM: "databricks-vscode",
                     BRICKS_UPSTREAM_VERSION: "1.0.0",
-                    DATABRICKS_CONFIG_PROFILE: "profile",
-                    DATABRICKS_CONFIG_FILE: undefined,
+                    DATABRICKS_HOST:
+                        "https://000000000000.00.azuredatabricks.net/",
+                    DATABRICKS_AUTH_TYPE: "metadata-service",
+                    DATABRICKS_METADATA_SERVICE_URL: "http://localhost:1234",
                     HOME: process.env.HOME,
                     PATH: process.env.PATH,
                     /* eslint-enable @typescript-eslint/naming-convention */
@@ -120,8 +126,10 @@ describe(__filename, () => {
                     BRICKS_ROOT: Uri.file("/path/to/local/workspace").fsPath,
                     BRICKS_UPSTREAM: "databricks-vscode",
                     BRICKS_UPSTREAM_VERSION: "1.0.0",
-                    DATABRICKS_CONFIG_PROFILE: "profile",
-                    DATABRICKS_CONFIG_FILE: undefined,
+                    DATABRICKS_HOST:
+                        "https://000000000000.00.azuredatabricks.net/",
+                    DATABRICKS_AUTH_TYPE: "metadata-service",
+                    DATABRICKS_METADATA_SERVICE_URL: "http://localhost:1234",
                     HOME: process.env.HOME,
                     PATH: process.env.PATH,
                     HTTP_PROXY: "http_proxy",
