@@ -6,6 +6,7 @@ import {promisify} from "node:util";
 import {withLogContext} from "@databricks/databricks-sdk/dist/logging";
 import {Loggers} from "../logger";
 import {Context, context} from "@databricks/databricks-sdk/dist/context";
+import {Cloud} from "../utils/constants";
 
 const execFile = promisify(execFileCb);
 
@@ -18,7 +19,7 @@ export interface ConfigEntry {
     name: string;
     host?: URL;
     accountId?: string;
-    cloud: "aws" | "azure" | "gcp";
+    cloud: Cloud;
     authType: string;
     valid: boolean;
 }
@@ -41,11 +42,8 @@ function getValidHost(host: string) {
 export class CliWrapper {
     constructor(private extensionContext: ExtensionContext) {}
 
-    getTestBricksCommand(): Command {
-        return {
-            command: this.extensionContext.asAbsolutePath("./bin/bricks"),
-            args: [],
-        };
+    get bricksPath(): string {
+        return this.extensionContext.asAbsolutePath("./bin/bricks");
     }
 
     /**
@@ -55,7 +53,6 @@ export class CliWrapper {
         syncDestination: SyncDestinationMapper,
         syncType: SyncType
     ): Command {
-        const command = this.extensionContext.asAbsolutePath("./bin/bricks");
         const args = [
             "sync",
             ".",
@@ -68,14 +65,14 @@ export class CliWrapper {
             args.push("--full");
         }
         if (workspaceConfigs.bricksVerboseMode) {
-            args.push("-v");
+            args.push("--log-level", "debug", "--log-file", "stderr");
         }
-        return {command, args};
+        return {command: this.bricksPath, args};
     }
 
     private getListProfilesCommand(): Command {
         return {
-            command: this.extensionContext.asAbsolutePath("./bin/bricks"),
+            command: this.bricksPath,
             args: ["auth", "profiles", "--skip-validate"],
         };
     }
@@ -136,16 +133,13 @@ export class CliWrapper {
 
     public async getBundleSchema(): Promise<string> {
         const execFile = promisify(execFileCb);
-        const {stdout} = await execFile(
-            this.extensionContext.asAbsolutePath("./bin/bricks"),
-            ["bundle", "schema"]
-        );
+        const {stdout} = await execFile(this.bricksPath, ["bundle", "schema"]);
         return stdout;
     }
 
     getAddProfileCommand(profile: string, host: URL): Command {
         return {
-            command: this.extensionContext.asAbsolutePath("./bin/bricks"),
+            command: this.bricksPath,
             args: [
                 "configure",
                 "--no-interactive",
