@@ -11,9 +11,11 @@ import {workspaceConfigs} from "../../vscode-objs/WorkspaceConfigs";
 const extensionVersion = require("../../../package.json")
     .version as ProductVersion;
 
-import {AzureCliCheck} from "../AzureCliCheck";
+import {AzureCliCheck} from "./AzureCliCheck";
+import {BricksCliCheck} from "./BricksCliCheck";
 
-export type AuthType = "azure-cli" | "google-id" | "oauth-u2m" | "profile";
+// TODO: Resolve this with SDK's AuthType.
+export type AuthType = "azure-cli" | "google-id" | "bricks-cli" | "profile";
 
 export abstract class AuthProvider {
     constructor(
@@ -34,7 +36,6 @@ export abstract class AuthProvider {
      */
     abstract describe(): string;
     abstract toJSON(): Record<string, unknown>;
-    abstract getEnvVars(): Record<string, string | undefined>;
 
     getWorkspaceClient(): WorkspaceClient {
         const config = this.getSdkConfig();
@@ -52,7 +53,10 @@ export abstract class AuthProvider {
 
     protected abstract getSdkConfig(): Config;
 
-    static fromJSON(json: Record<string, any>): AuthProvider {
+    static fromJSON(
+        json: Record<string, any>,
+        bricksPath: string
+    ): AuthProvider {
         const host =
             json.host instanceof URL
                 ? json.host
@@ -72,6 +76,9 @@ export abstract class AuthProvider {
                     json.tenantId,
                     json.appId
                 );
+
+            case "bricks-cli":
+                return new BricksCliAuthProvider(host, bricksPath);
 
             case "profile":
                 if (!json.profile) {
@@ -102,15 +109,6 @@ export class ProfileAuthProvider extends AuthProvider {
         };
     }
 
-    getEnvVars(): Record<string, string | undefined> {
-        return {
-            DATABRICKS_CONFIG_PROFILE: this.profile,
-            DATABRICKS_CONFIG_FILE:
-                workspaceConfigs.databrickscfgLocation ??
-                process.env.DATABRICKS_CONFIG_FILE,
-        };
-    }
-
     getSdkConfig(): Config {
         return new Config({
             profile: this.profile,
@@ -119,6 +117,37 @@ export class ProfileAuthProvider extends AuthProvider {
                 process.env.DATABRICKS_CONFIG_FILE,
             env: {},
         });
+    }
+}
+
+export class BricksCliAuthProvider extends AuthProvider {
+    constructor(host: URL, readonly bricksPath: string) {
+        super(host, "bricks-cli");
+    }
+
+    describe(): string {
+        return "OAuth U2M";
+    }
+
+    toJSON(): Record<string, unknown> {
+        return {
+            host: this.host.toString(),
+            authType: this.authType,
+            bricksPath: this.bricksPath,
+        };
+    }
+
+    getSdkConfig(): Config {
+        return new Config({
+            host: this.host.toString(),
+            authType: "bricks-cli",
+            bricksCliPath: this.bricksPath,
+        });
+    }
+
+    async check(silent: boolean): Promise<boolean> {
+        const bricksCliCheck = new BricksCliCheck(this);
+        return bricksCliCheck.check(silent);
     }
 }
 
@@ -151,13 +180,6 @@ export class AzureCliAuthProvider extends AuthProvider {
             authType: this.authType,
             tenantId: this.tenantId,
             appId: this.appId,
-        };
-    }
-
-    getEnvVars(): Record<string, string | undefined> {
-        return {
-            DATABRICKS_HOST: this.host.toString(),
-            DATABRICKS_AUTH_TYPE: this.authType,
         };
     }
 
