@@ -5,73 +5,15 @@ import {
 } from "@databricks/databricks-sdk";
 import {NamedLogger} from "@databricks/databricks-sdk/dist/logging";
 import {commands, Disposable, Uri, window} from "vscode";
-import {Loggers} from "../logger";
-import {AzureCliAuthProvider} from "./auth/AuthProvider";
+import {Loggers} from "../../logger";
+import {AzureCliAuthProvider} from "./AuthProvider";
+import {orchestrate, OrchestrationLoopError, Step} from "./orchestrate";
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
-const extensionVersion = require("../../package.json")
+const extensionVersion = require("../../../package.json")
     .version as ProductVersion;
 
 type AzureCloud = "AzureCloud" | "AzureChinaCloud" | "AzureUSGovernment";
-
-export type Step<S, N> = () => Promise<
-    SuccessResult<S> | NextResult<N> | ErrorResult
->;
-
-type StepResult<S, N> = SuccessResult<S> | NextResult<N> | ErrorResult;
-
-export interface SuccessResult<T> {
-    type: "success";
-    result: T;
-}
-
-export interface NextResult<T> {
-    type: "next";
-    next: T;
-}
-
-export interface ErrorResult {
-    type: "error";
-    error: Error;
-}
-
-class OrchestrationLoopError extends Error {
-    constructor() {
-        super("OrchestrationLoopError");
-    }
-}
-
-async function orchestrate<S, KEYS extends string>(
-    steps: Record<KEYS, Step<S, KEYS>>,
-    start: KEYS,
-    maxSteps = 20,
-    logger?: NamedLogger
-): Promise<S> {
-    let counter = 0;
-
-    let step: KEYS | undefined = start;
-    while (step && steps[step]) {
-        counter += 1;
-        if (counter > maxSteps) {
-            throw new OrchestrationLoopError();
-        }
-        const result: StepResult<S, KEYS> = await steps[step]();
-        logger?.info(`Azure CLI check: ${step}`, result);
-
-        if (result.type === "error") {
-            throw result.error;
-        }
-
-        if (result.type === "success") {
-            return result.result;
-        }
-
-        if (result.next) {
-            step = result.next;
-        }
-    }
-    throw new Error("Missing return step");
-}
 
 type AzureStepName =
     | "tryLogin"
