@@ -36,9 +36,14 @@ import {generateBundleSchema} from "./bundle/GenerateBundle";
 import {CustomWhenContext} from "./vscode-objs/CustomWhenContext";
 import {WorkspaceStateManager} from "./vscode-objs/WorkspaceState";
 import path from "node:path";
+import {MetadataServiceManager} from "./configuration/auth/MetadataServiceManager";
 import {FeatureManager} from "./feature-manager/FeatureManager";
 import {DbConnectAccessVerifier} from "./language/DbConnectAccessVerifier";
 import {MsPythonExtensionWrapper} from "./language/MsPythonExtensionWrapper";
+import {DatabricksEnvFileManager} from "./file-managers/DatabricksEnvFileManager";
+import {Telemetry, toUserMetadata} from "./telemetry";
+import "./telemetry/commandExtensions";
+import {Events, Metadata} from "./telemetry/constants";
 
 export async function activate(
     context: ExtensionContext
@@ -73,6 +78,8 @@ export async function activate(
         return undefined;
     }
 
+    const workspaceStateManager = new WorkspaceStateManager(context);
+
     // Add the bricks binary to the PATH environment variable in terminals
     context.environmentVariableCollection.persistent = true;
     context.environmentVariableCollection.prepend(
@@ -84,6 +91,8 @@ export async function activate(
     if (workspaceConfigs.loggingEnabled) {
         loggerManager.initLoggers();
     }
+
+    const telemetry = Telemetry.createDefault();
 
     const packageMetadata = await PackageJsonUtils.getMetadata(context);
     NamedLogger.getOrCreate(Loggers.Extension).debug("Metadata", {
@@ -103,10 +112,9 @@ export async function activate(
 
     const pythonExtensionWrapper = new MsPythonExtensionWrapper(
         pythonExtension,
-        workspace.workspaceFolders[0].uri
+        workspace.workspaceFolders[0].uri,
+        workspaceStateManager
     );
-
-    const workspaceStateManager = new WorkspaceStateManager(context);
 
     const configureAutocomplete = new ConfigureAutocomplete(
         context,
@@ -116,7 +124,7 @@ export async function activate(
     );
     context.subscriptions.push(
         configureAutocomplete,
-        commands.registerCommand(
+        telemetry.registerCommand(
             "databricks.autocomplete.configure",
             configureAutocomplete.configureCommand,
             configureAutocomplete
@@ -124,7 +132,7 @@ export async function activate(
     );
 
     context.subscriptions.push(
-        commands.registerCommand(
+        telemetry.registerCommand(
             "databricks.logs.openFolder",
             loggerManager.openLogFolder,
             loggerManager
@@ -132,8 +140,24 @@ export async function activate(
     );
 
     const cli = new CliWrapper(context);
+
     // Configuration group
     const connectionManager = new ConnectionManager(cli);
+    context.subscriptions.push(
+        connectionManager.onDidChangeState(async () => {
+            telemetry.setMetadata(
+                Metadata.USER,
+                await toUserMetadata(connectionManager.databricksWorkspace)
+            );
+            telemetry.recordEvent(Events.CONNECTION_STATE_CHANGED, {
+                newState: connectionManager.state,
+            });
+        })
+    );
+    const metadataServiceManager = new MetadataServiceManager(
+        connectionManager
+    );
+    await metadataServiceManager.listen();
 
     const workspaceFsDataProvider = new WorkspaceFsDataProvider(
         connectionManager
@@ -146,16 +170,17 @@ export async function activate(
     );
 
     context.subscriptions.push(
+        metadataServiceManager,
         window.registerTreeDataProvider(
             "workspaceFsView",
             workspaceFsDataProvider
         ),
-        commands.registerCommand(
+        telemetry.registerCommand(
             "databricks.wsfs.refresh",
             workspaceFsCommands.refresh,
             workspaceFsCommands
         ),
-        commands.registerCommand(
+        telemetry.registerCommand(
             "databricks.wsfs.createFolder",
             workspaceFsCommands.createFolder,
             workspaceFsCommands
@@ -187,42 +212,42 @@ export async function activate(
             "configurationView",
             configurationDataProvider
         ),
-        commands.registerCommand(
+        telemetry.registerCommand(
             "databricks.connection.logout",
             connectionCommands.logoutCommand(),
             connectionCommands
         ),
-        commands.registerCommand(
+        telemetry.registerCommand(
             "databricks.connection.configureWorkspace",
             connectionCommands.configureWorkspaceCommand(),
             connectionCommands
         ),
-        commands.registerCommand(
+        telemetry.registerCommand(
             "databricks.connection.openDatabricksConfigFile",
             connectionCommands.openDatabricksConfigFileCommand(),
             connectionCommands
         ),
-        commands.registerCommand(
+        telemetry.registerCommand(
             "databricks.connection.attachCluster",
             connectionCommands.attachClusterCommand(),
             connectionCommands
         ),
-        commands.registerCommand(
+        telemetry.registerCommand(
             "databricks.connection.attachClusterQuickPick",
             connectionCommands.attachClusterQuickPickCommand(),
             connectionCommands
         ),
-        commands.registerCommand(
+        telemetry.registerCommand(
             "databricks.connection.detachCluster",
             connectionCommands.detachClusterCommand(),
             connectionCommands
         ),
-        commands.registerCommand(
+        telemetry.registerCommand(
             "databricks.connection.attachSyncDestination",
             connectionCommands.attachSyncDestinationCommand(),
             connectionCommands
         ),
-        commands.registerCommand(
+        telemetry.registerCommand(
             "databricks.connection.detachSyncDestination",
             connectionCommands.detachWorkspaceCommand(),
             connectionCommands
@@ -252,12 +277,12 @@ export async function activate(
     );
 
     context.subscriptions.push(
-        commands.registerCommand(
+        telemetry.registerCommand(
             "databricks.run.runEditorContents",
             runCommands.runEditorContentsCommand(),
             runCommands
         ),
-        commands.registerCommand(
+        telemetry.registerCommand(
             "databricks.run.runEditorContentsAsWorkflow",
             runCommands.runEditorContentsAsWorkflowCommand(),
             runCommands
@@ -283,32 +308,32 @@ export async function activate(
         clusterTreeDataProvider,
         window.registerTreeDataProvider("clusterView", clusterTreeDataProvider),
 
-        commands.registerCommand(
+        telemetry.registerCommand(
             "databricks.cluster.refresh",
             clusterCommands.refreshCommand(),
             clusterCommands
         ),
-        commands.registerCommand(
+        telemetry.registerCommand(
             "databricks.cluster.filterByAll",
             clusterCommands.filterCommand("ALL"),
             clusterCommands
         ),
-        commands.registerCommand(
+        telemetry.registerCommand(
             "databricks.cluster.filterByRunning",
             clusterCommands.filterCommand("RUNNING"),
             clusterCommands
         ),
-        commands.registerCommand(
+        telemetry.registerCommand(
             "databricks.cluster.filterByMe",
             clusterCommands.filterCommand("ME"),
             clusterCommands
         ),
-        commands.registerCommand(
+        telemetry.registerCommand(
             "databricks.cluster.start",
             clusterCommands.startClusterCommand,
             clusterCommands
         ),
-        commands.registerCommand(
+        telemetry.registerCommand(
             "databricks.cluster.stop",
             clusterCommands.stopClusterCommand,
             clusterCommands
@@ -318,17 +343,17 @@ export async function activate(
     // Sync
     const syncCommands = new SyncCommands(synchronizer);
     context.subscriptions.push(
-        commands.registerCommand(
+        telemetry.registerCommand(
             "databricks.sync.start",
             syncCommands.startCommand("incremental"),
             syncCommands
         ),
-        commands.registerCommand(
+        telemetry.registerCommand(
             "databricks.sync.startFull",
             syncCommands.startCommand("full"),
             syncCommands
         ),
-        commands.registerCommand(
+        telemetry.registerCommand(
             "databricks.sync.stop",
             syncCommands.stopCommand(),
             syncCommands
@@ -336,13 +361,17 @@ export async function activate(
     );
 
     context.subscriptions.push(
-        new ProjectConfigFileWatcher(connectionManager, workspace.rootPath)
+        new ProjectConfigFileWatcher(
+            connectionManager,
+            workspace.rootPath!,
+            cli.bricksPath
+        )
     );
 
     // Quickstart
     const quickstartCommands = new QuickstartCommands(context);
     context.subscriptions.push(
-        commands.registerCommand(
+        telemetry.registerCommand(
             "databricks.quickstart.open",
             quickstartCommands.openQuickstartCommand(),
             quickstartCommands
@@ -356,11 +385,16 @@ export async function activate(
     // Utils
     const utilCommands = new UtilsCommands.UtilsCommands();
     context.subscriptions.push(
-        commands.registerCommand(
+        telemetry.registerCommand(
             "databricks.utils.openExternal",
             utilCommands.openExternalCommand(),
             utilCommands
-        )
+        ),
+        telemetry.registerCommand("databricks.call", (fn) => {
+            if (fn) {
+                fn();
+            }
+        })
     );
 
     // generate a json schema for bundle root and load a custom provider into
@@ -375,49 +409,40 @@ export async function activate(
     const featureManager = new FeatureManager<"debugging.dbconnect">([
         "debugging.dbconnect",
     ]);
-    const dbConnectAccessVerifier = new DbConnectAccessVerifier(
-        connectionManager,
-        pythonExtensionWrapper,
-        workspaceStateManager
-    );
     featureManager.registerFeature(
         "debugging.dbconnect",
-        dbConnectAccessVerifier
+        () =>
+            new DbConnectAccessVerifier(
+                connectionManager,
+                pythonExtensionWrapper,
+                workspaceStateManager,
+                context
+            )
+    );
+    const databricksEnvFileManager = new DatabricksEnvFileManager(
+        workspace.workspaceFolders[0].uri,
+        featureManager,
+        connectionManager,
+        context
     );
     context.subscriptions.push(
-        dbConnectAccessVerifier,
-        featureManager.onDidChangeState(
-            "debugging.dbconnect",
-            async (state) => {
-                if (state.avaliable) {
-                    window.showInformationMessage("Db Connect is enabled");
-                    return;
-                }
-                if (state.reason) {
-                    window.showErrorMessage(
-                        `DB Connect is disabled because ${state.reason}`
-                    );
-                }
-                if (state.action) {
-                    await state.action();
-                }
+        databricksEnvFileManager,
+        databricksEnvFileManager.onDidChangeEnvironmentVariables(() => {
+            if (workspace.notebookDocuments.length) {
+                window.showInformationMessage(
+                    "Environment variables have changed. Restart all jupyter kernels to pickup the latest environment variables."
+                );
             }
-        )
+        })
     );
     featureManager.isEnabled("debugging.dbconnect");
 
-    context.subscriptions.push(
-        commands.registerCommand(
-            "databricks.test.pipFreeze",
-            dbConnectAccessVerifier.checkDbConnectInstall,
-            dbConnectAccessVerifier
-        )
-    );
     connectionManager.login(false).catch((e) => {
         NamedLogger.getOrCreate(Loggers.Extension).error("Login error", e);
     });
 
     CustomWhenContext.setActivated(true);
+    telemetry.recordEvent(Events.EXTENSION_ACTIVATED);
     return {
         connectionManager: connectionManager,
     };
