@@ -1,31 +1,28 @@
-import {Cluster} from "@databricks/databricks-sdk";
-import {window} from "vscode";
+import {Cluster, logging} from "@databricks/databricks-sdk";
+import {window, ExtensionContext, commands} from "vscode";
+import * as path from "node:path";
 import {ConnectionManager} from "../configuration/ConnectionManager";
 import {MultiStepAccessVerifier} from "../feature-manager/MultiStepAccessVerfier";
 import {WorkspaceStateManager} from "../vscode-objs/WorkspaceState";
 import {MsPythonExtensionWrapper} from "./MsPythonExtensionWrapper";
+import {Loggers} from "../logger";
+import {Context, context} from "@databricks/databricks-sdk/dist/context";
 
 export class DbConnectAccessVerifier extends MultiStepAccessVerifier {
     constructor(
         private readonly connectionManager: ConnectionManager,
         private readonly pythonExtension: MsPythonExtensionWrapper,
-        private readonly workspaceState: WorkspaceStateManager
+        private readonly workspaceState: WorkspaceStateManager,
+        private readonly context: ExtensionContext
     ) {
-        super([
-            "checkClusterVersion",
-            "checkClusterHasUc",
-            "checkWorkspaceHasUc",
-            "checkDbConnectInstall",
-        ]);
+        super(["checkCluster", "checkWorkspaceHasUc", "checkDbConnectInstall"]);
         this.disposables.push(
             this.connectionManager.onDidChangeCluster((cluster) => {
                 if (this.connectionManager.state !== "CONNECTED") {
                     return;
                 }
                 const steps = {
-                    checkClusterVersion: () =>
-                        this.checkClusterVersion(cluster),
-                    checkClusterHasUc: () => this.checkClusterHasUc(cluster),
+                    checkCluster: () => this.checkCluster(cluster),
                 };
                 this.runSteps(steps);
             }, this),
@@ -45,44 +42,65 @@ export class DbConnectAccessVerifier extends MultiStepAccessVerifier {
         );
     }
 
-    async checkClusterVersion(cluster?: Cluster) {
+    promptForAttachingCluster(msg: string) {
+        return async () => {
+            const choice = await window.showInformationMessage(
+                msg,
+                "Attach Cluster",
+                "Ignore"
+            );
+            switch (choice) {
+                case "Attach Cluster":
+                    commands.executeCommand(
+                        "databricks.connection.attachClusterQuickPick"
+                    );
+            }
+        };
+    }
+
+    async checkCluster(cluster?: Cluster) {
         if (cluster === undefined) {
             return this.rejectStep(
-                "checkClusterVersion",
-                "No cluster attached"
+                "checkCluster",
+                "No cluster attached",
+                this.promptForAttachingCluster(
+                    "No cluster is attached. Please attach a cluster."
+                )
             );
         }
         await this.connectionManager.waitForConnect();
 
         const dbrVersionParts = cluster?.dbrVersion;
         if (
-            dbrVersionParts &&
-            (dbrVersionParts[0] === "x" || dbrVersionParts[0] >= 13)
+            !dbrVersionParts ||
+            (dbrVersionParts[0] !== "x" && dbrVersionParts[0] < 13)
         ) {
-            return this.acceptStep("checkClusterVersion");
-        }
-        return this.rejectStep(
-            "checkClusterVersion",
-            `Cluster dbr is less than 13.0.0`
-        );
-    }
-
-    async checkClusterHasUc(cluster?: Cluster) {
-        if (cluster === undefined) {
-            return this.rejectStep("checkClusterHasUc", "No cluster attached");
-        }
-        this.connectionManager.waitForConnect();
-
-        if (cluster.isUc()) {
             return this.rejectStep(
-                "checkClusterHasUc",
-                `Cluster doesn't have UC enabled (access mode should be "Single User" or "Shared")`
+                "checkCluster",
+                `DB Connect V2 requires cluster DBR >= 13.0.0.`,
+                this.promptForAttachingCluster(
+                    `DB Connect V2 requires cluster DBR >= 13.0.0. Currently it is ${dbrVersionParts.join(
+                        "."
+                    )}. Please attach a new cluster.`
+                )
             );
         }
-        return this.acceptStep("checkClusterHasUc");
+        if (!cluster.isUc()) {
+            return this.rejectStep(
+                "checkCluster",
+                `Cluster doesn't have UC enabled.`,
+                this.promptForAttachingCluster(
+                    `Access mode should be "Single User" or "Shared"). Currently it is ${
+                        cluster.accessMode ?? "custom"
+                    }. Please attach a new cluster.`
+                )
+            );
+        }
+        return this.acceptStep("checkCluster");
     }
 
-    async checkWorkspaceHasUc() {
+    @logging.withLogContext(Loggers.Extension)
+    async checkWorkspaceHasUc(@context ctx?: Context) {
         this.connectionManager.waitForConnect();
         try {
             const catalogList =
@@ -90,11 +108,12 @@ export class DbConnectAccessVerifier extends MultiStepAccessVerifier {
             if (!catalogList?.catalogs?.length) {
                 return this.rejectStep(
                     "checkWorkspaceHasUc",
-                    "Can't find any catalogs with read permissions"
+                    "No catalogues with read permission were found. Please enable UC for the workspace."
                 );
             }
         } catch (e: unknown) {
             if (e instanceof Error) {
+                ctx?.logger?.error("Error while searching for catalogues", e);
                 return this.rejectStep("checkWorkspaceHasUc", e.message);
             }
         }
@@ -113,8 +132,16 @@ export class DbConnectAccessVerifier extends MultiStepAccessVerifier {
             return;
         }
 
+        const wheelPath = this.context.asAbsolutePath(
+            path.join(
+                "resources",
+                "python",
+                "databricks_connect-13.0.0-py2.py3-none-any.whl"
+            )
+        );
+
         const choice = await window.showInformationMessage(
-            "Do you want to install dbconnectV2 in the current environment?",
+            "Do you want to install databricks-connect in the current environment?",
             "Yes",
             "Ignore",
             "Ignore for environment",
@@ -128,7 +155,7 @@ export class DbConnectAccessVerifier extends MultiStepAccessVerifier {
                         "pyspark"
                     );
                     await this.pythonExtension.installPackageInEnvironment(
-                        "pyspark"
+                        wheelPath
                     );
                     this.checkDbConnectInstall();
                 } catch (e: unknown) {
@@ -155,19 +182,37 @@ export class DbConnectAccessVerifier extends MultiStepAccessVerifier {
         if (!executable) {
             return this.rejectStep(
                 "checkDbConnectInstall",
-                "No python executable found"
+                "No python executable found."
+            );
+        }
+        const env = await this.pythonExtension.pythonEnvironment;
+        if (
+            env &&
+            env.version &&
+            !(
+                env.version.major > 3 ||
+                (env.version.major === 3 && env.version.minor >= 9)
+            )
+        ) {
+            return this.rejectStep(
+                "checkDbConnectInstall",
+                `DB Connect V2 requires python >= 3.9.0. Current version is ${[
+                    env.version.major,
+                    env.version.minor,
+                    env.version.micro,
+                ].join(".")}`
             );
         }
         try {
             const exists = await this.pythonExtension.findPackageInEnvironment(
-                "pyspark"
+                "databricks-connect"
             );
             if (exists) {
                 return this.acceptStep("checkDbConnectInstall");
             }
             return this.rejectStep(
                 "checkDbConnectInstall",
-                "Dbconnect package not installed in the current environment",
+                "databricks-connect package is not installed in the current environment",
                 () => this.showDbConnectInstallPrompt()
             );
         } catch (e: unknown) {
@@ -182,10 +227,8 @@ export class DbConnectAccessVerifier extends MultiStepAccessVerifier {
     override async check() {
         await this.connectionManager.waitForConnect();
         const steps = {
-            checkClusterVersion: () =>
-                this.checkClusterVersion(this.connectionManager.cluster),
-            checkClusterHasUc: () =>
-                this.checkClusterHasUc(this.connectionManager.cluster),
+            checkCluster: () =>
+                this.checkCluster(this.connectionManager.cluster),
             checkWorkspaceHasUc: () => this.checkWorkspaceHasUc(),
             checkDbConnectInstall: () => this.checkDbConnectInstall(),
         };
